var cleaning and removing extraneous serial init.

+++ ArduinoFocuser/ArduinoFocuser/PushButton.h
@@ -27,8 +27,8 @@ private:
 	bool pushed = false, held=false, lit=false, blink=false;
 	long lastButtonPushTime = 0;
 	int heldThreshold = 1 * 1000; // 2 Seconds
-	int debounceTime = 30; // 50 milliseconds
-	int defaultBrightness = 128;
+	uint8_t debounceTime = 30; // 50 milliseconds
+	uint8_t defaultBrightness = 128;
 };
 
 #endif
